Renamed `core_function.rs` to `main_loop.rs`
