+++ src/core/main_loop.rs
